Zoo Room Art Implementation
Added in Cat, Dog, Cloud, and Duck gifs along with assorted frames from Cat and Dog to Animations folder, added in various buttons into the Zoo Room assets folder and changed out the still frames of the animals with the buttons, and changed out Rizzy and Cowboy Steve. No changes have been made to the scene otherwise.

Looking for Feedback on changes.