Folder Organization: Rooms
Went through the entirety of the art assets folder of the rooms and main menu, leaving the characters and minigames folder alone. Made an Archive folder for all the old assets in the main assets folder, moved a room folder with all the room assets to the main folder with all requisite folders required. Needs further review if job done correctly.